Spike on interface introspection.
The idea here is to enable the interface to be tested while still only being defined in one place.

I've left in the failing tests to show what is being tested.

This isn't robust yet. It won't handled circular references (maybe
that's ok). And it won't handle sources that are just symbols, not Ask
objects.