Feature: Interface object for Passive Domain objects
This Interface object is instantiated with another Passive Domain value
object. It offers the methods:

calls - a hash of the methods that the source object calls on its source
mapped to the Only object that limit what valid responsed must be.

responds_to - an array of the public methods made available by the source
object, excluding all the standard ones.

Two interfaces are compatible if calls.keys of one is a subset of the
responds_to of the other.